requirements.txt
appdirs==1.4.0
dnspython==1.15.0
numpy==1.12.0
packaging==16.8
pandas==0.19.2
pyparsing==2.1.10
python-dateutil==2.6.0
pytz==2016.10
six==1.10.0
